+++ src/index.html
@@ -17,7 +17,7 @@
     <div class="page2">
       
       <br>
-      <input type="search" value=""> <input type="submit" id="search" value="Buscar">
+      <input type="search" id="search" value=""> <input type="submit" id="submitButton" value="Buscar">
       <p>Aquí puedes buscar tu pokémon favorito</p>
      <h1>Tipos</h1>
       <button class="button1" id="steelButton" value="steel">Acero</button>
@@ -42,8 +42,12 @@
 
     <div class="page3">
 
-    <p>3ra página</p>
-
+        
+      <h1 id="title" value=""><h1>
+        
+      <figcaption class="info" id="info" value=""  ></figcaption>
+          
+        
     </div>
     <script src="main.js" type="module"></script>
   </body>
