+++ src/main.js
@@ -1,8 +1,8 @@
 import { example } from './data.js';
-// import data from './data/lol/lol.js';
-import data from './data/pokemon/pokemon.js';
-// import data from './data/rickandmorty/rickandmorty.js';
-//declarar constantes
+
+// import data from './data/pokemon/pokemon.js';
+
+
 
 
 //pasar de 1ra página a 2da página
@@ -16,6 +16,7 @@ const buttonPage2 = () => {
 buttonClick.addEventListener("click", buttonPage2);
 
 //pasar de 2da página a 3ra página
+const sbmt =document.getElementById("submitButton")
 const button1 =document.getElementById("steelButton")
 const button2 =document.getElementById("waterButton")
 const button3 =document.getElementById("bugButton")
@@ -34,13 +35,12 @@ const button15 =document.getElementById("groundButton")
 const button16 =document.getElementById("poisonButton")
 const button17 =document.getElementById("flyingButton")
 
-
-const buttonPage3 = (e) => {
-     let buttonValue=e.target.value     
+const buttonPage3 = () => { 
     document.querySelector(".page2").style.display = "none";
     document.querySelector(".page3").style.display = "block";
 }
 
+sbmt.addEventListener("click", buttonPage3)
 button1.addEventListener("click", buttonPage3);
 button2.addEventListener("click", buttonPage3);
 button3.addEventListener("click", buttonPage3);
@@ -61,49 +61,5 @@ button17.addEventListener("click", buttonPage3);
 
 
 
-//seleccinar tipo de pokemón
-const steelPokemon = document.getElementById("steelButton");
-const waterPokemon = document.getElementById("waterButton");
-const bugPokemon = document.getElementById("bugButton");
-const electricPokemon = document.getElementById("electricButton");
-const ghostPokemon = document.getElementById("ghostButton");
-const firePokemon = document.getElementById("fireButton");
-const fairyPokemon = document.getElementById("fairyButton");
-const icePokemon = document.getElementById("iceButton");
-const fightingPokemon = document.getElementById("fightingButton");
-const normalPokemon = document.getElementById("normalButton");
-const grassPokemon = document.getElementById("grassButton");
-const psychicPokemon = document.getElementById("psychicButton");
-const rockPokemon = document.getElementById("rockButton");
-const darkPokemon = document.getElementById("darkButton");
-const groundPokemon = document.getElementById("groundButton");
-const poisonPokemon = document.getElementById("poisonButton");
-const flyingPokemon = document.getElementById("flyingButton");
-
-const typeFunction = (e) => {
-     let buttonValue=e.target.value
-     let pokemon1= data.pokemon.filter(x => x.type.includes(buttonValue));
-console.log(buttonValue)    
-console.log(pokemon1)
-}
-
-steelPokemon.addEventListener("click", typeFunction);
-waterPokemon.addEventListener("click", typeFunction);
-bugPokemon.addEventListener("click", typeFunction);
-electricPokemon.addEventListener("click", typeFunction);
-ghostPokemon.addEventListener("click", typeFunction);
-firePokemon.addEventListener("click", typeFunction);
-fairyPokemon.addEventListener("click", typeFunction);
-icePokemon.addEventListener("click", typeFunction);
-fightingPokemon.addEventListener("click", typeFunction);
-normalPokemon.addEventListener("click", typeFunction);
-grassPokemon.addEventListener("click", typeFunction);
-psychicPokemon.addEventListener("click", typeFunction);
-rockPokemon.addEventListener("click", typeFunction);
-darkPokemon.addEventListener("click", typeFunction);
-groundPokemon.addEventListener("click", typeFunction);
-poisonPokemon.addEventListener("click", typeFunction);
-flyingPokemon.addEventListener("click", typeFunction);
 
 // console.log(example, data);
-
